Moved tests.
