+++ tests/it/io/ipc/mod.rs
@@ -1,0 +1,5 @@
+mod common;
+mod read;
+mod write;
+
+pub use common::read_gzip_json;
