+++ tests/it/io/ipc/read/mod.rs
@@ -1,0 +1,2 @@
+mod file;
+mod stream;
