+++ tests/it/io/ipc/write/mod.rs
@@ -1,0 +1,2 @@
+mod file;
+mod stream;
