+++ tests/it/io/json/mod.rs
@@ -1,1 +1,2 @@
 mod read;
+mod write;
