+++ tests/it/io/mod.rs
@@ -3,3 +3,9 @@ mod print;
 
 #[cfg(feature = "io_json")]
 mod json;
+
+#[cfg(feature = "io_ipc")]
+mod ipc;
+
+#[cfg(feature = "io_parquet")]
+mod parquet;
