+++ tests/it/main.rs
@@ -4,3 +4,4 @@ mod buffer;
 mod ffi;
 
 mod io;
+mod test_util;
